requirements.txt
streamlit
google-cloud-bigquery
pandas
db_dtypes
gcsfs
st-files-connection